chore(BK): Add DRY_RUN to avoid upload

+++ scripts/upload_re2_artifacts.sh
@@ -10,4 +10,10 @@ ARTIFACT_DIST_DIR="./workdir_re2/dist"
 list_shasums_in_folder $ARTIFACT_DIST_DIR
 
 echo "--- Uploading build artifacts"
-gsutil cp -r "$ARTIFACT_DIST_DIR/*" gs://$BUCKET_NAME/$ARTIFACT_BASE_PATH
+echo "--- Uploading build artifacts"
+if [[ "${DRY_RUN:-}" == "1" || "${DRY_RUN:-}" == "true" ]]; then
+  echo "Dry run enabled, skipping upload"
+  exit 0
+else
+  gsutil cp -r "$ARTIFACT_DIST_DIR/*" gs://$BUCKET_NAME/$ARTIFACT_BASE_PATH
+fi
